feat: Implement Google Sign-In with One Tap
