+++ package.json
@@ -10,17 +10,18 @@
   },
   "dependencies": {
     "@emailjs/browser": "^4.4.1",
-    "@firebase-oss/ui-core": "7.0.2-beta",
-    "@firebase-oss/ui-react": "7.0.2-beta",
-    "@firebase-oss/ui-styles": "7.0.2-beta",
     "@googleworkspace/meet-addons": "^1.2.0",
+    "@mui/icons-material": "^7.3.6",
+    "@mui/material": "^7.3.6",
+    "@tanstack/react-query": "^5.90.12",
     "axios": "^1.1.3",
     "firebase": "^11.0.0",
     "next": "16.0.10",
     "react": "19.2.3",
     "react-dom": "19.2.3",
     "react-firebase-hooks": "^5.0.3",
-    "tailwind-merge": "^1.8.0"
+    "tailwind-merge": "^1.8.0",
+    "zustand": "^5.0.9"
   },
   "devDependencies": {
     "@types/aos": "^3.0.4",
