+++ pages/_app.tsx
@@ -1,6 +1,5 @@
 import "aos/dist/aos.css";
 import "../styles/globals.css";
-import "../styles/base.css";
 
 import AOS from "aos";
 import type { AppProps } from "next/app";
