+++ src/components/GoogleButton.tsx
@@ -1,94 +1,74 @@
-import { useSignInWithProvider } from "@firebase-oss/ui-react";
-
-type Props = {};
-import { useSignInWithGoogle } from "react-firebase-hooks/auth";
+import { useEffect } from "react";
+import { useAuthState } from "react-firebase-hooks/auth";
 import { auth } from "../firebase/init";
-import { getAuth, signInWithPopup, GoogleAuthProvider } from "firebase/auth";
+import { signOut } from "firebase/auth";
+import signInWithGoogleIdToken from "../firebase/authHelpers";
 
-const provider = new GoogleAuthProvider();
+declare const google: any;
 
-const GoogleButton = (props: Props) => {
-  // const signInWithProvider = useSignInWithGoogle(auth);
-  const [signInWithGoogle, googleUser, googleLoading, googleError] =
-    useSignInWithGoogle(auth);
+const GoogleButton = () => {
+  const [user] = useAuthState(auth);
 
-  const signIn = async () => {
-    // const details = await signInWithGoogle();
-    // console.log("Signed in with Google:", details);
-    // console.log(googleUser);
+  useEffect(() => {
+    function GoogleLogin() {
+      if (typeof window === "undefined" || typeof google === "undefined")
+        return;
 
-    signInWithPopup(auth, provider)
-      .then((result) => {
-        // This gives you a Google Access Token. You can use it to access the Google API.
-        const credential = GoogleAuthProvider.credentialFromResult(result);
-        const token = credential ? credential.accessToken : null;
-        // The signed-in user info.
-        const user = result.user;
-        console.log({
-          credential,
-          token,
-          user,
-        });
+      // If user is signed in, cancel any One Tap prompts and don't initialize again
+      if (user) {
+        try {
+          google.accounts.id.cancel();
+        } catch (e) {
+          // ignore if cancel not supported
+        }
+        return;
+      }
 
-        // IdP data available using getAdditionalUserInfo(result)
-        // ...
-      })
-      .catch((error) => {
-        // Handle Errors here.
-        const errorCode = error.code;
-        const errorMessage = error.message;
-        // The email of the user's account used.
-        const email = error.customData.email;
-        // The AuthCredential type that was used.
-        const credential = GoogleAuthProvider.credentialFromError(error);
-        // ...
-        console.log({
-          errorCode,
-          errorMessage,
-          email,
-          credential,
-        });
+      google.accounts.id.initialize({
+        client_id: process.env.NEXT_PUBLIC_GOOGLE_CLIENT_ID!,
+        callback: async function (response: any) {
+          // response.credential is the Google ID token (JWT)
+          try {
+            const firebaseUser = await signInWithGoogleIdToken(
+              response.credential
+            );
+            console.log("Signed into Firebase via ID token:", firebaseUser);
+          } catch (err) {
+            console.error("Failed to sign in with ID token:", err);
+          }
+        },
       });
-  };
-  console.log({
-    googleUser,
-    googleError,
-  });
 
-  return (
-    <button
-      className="fui-button fui-button--secondary fui-provider__button"
-      type="button"
-      data-provider="google.com"
-      onClick={signIn}
-    >
-      <svg
-        xmlns="http://www.w3.org/2000/svg"
-        width="1em"
-        height="1em"
-        viewBox="0 0 48 48"
-        className="fui-provider__icon"
+      google.accounts.id.renderButton(
+        document.getElementById("googleSignInButton"),
+        { theme: "outline", size: "large" } // customization attributes
+      );
+      google.accounts.id.prompt(); // also display the One Tap dialog
+    }
+    GoogleLogin();
+  }, [user]);
+
+  const handleSignOut = async () => {
+    try {
+      await signOut(auth);
+      console.log("Signed out");
+    } catch (e) {
+      console.error("Sign out failed:", e);
+    }
+  };
+
+  if (user) {
+    return (
+      <button
+        className="fui-button fui-button--secondary"
+        onClick={handleSignOut}
       >
-        <path
-          fill="#FFC107"
-          d="M43.611 20.083H42V20H24v8h11.303c-1.649 4.657-6.08 8-11.303 8-6.627 0-12-5.373-12-12s5.373-12 12-12c3.059 0 5.842 1.154 7.961 3.039l5.657-5.657C34.046 6.053 29.268 4 24 4 12.955 4 4 12.955 4 24s8.955 20 20 20 20-8.955 20-20c0-1.341-.138-2.65-.389-3.917"
-        ></path>
-        <path
-          fill="#FF3D00"
-          d="m6.306 14.691 6.571 4.819C14.655 15.108 18.961 12 24 12c3.059 0 5.842 1.154 7.961 3.039l5.657-5.657C34.046 6.053 29.268 4 24 4 16.318 4 9.656 8.337 6.306 14.691"
-        ></path>
-        <path
-          fill="#4CAF50"
-          d="M24 44c5.166 0 9.86-1.977 13.409-5.192l-6.19-5.238A11.9 11.9 0 0 1 24 36c-5.202 0-9.619-3.317-11.283-7.946l-6.522 5.025C9.505 39.556 16.227 44 24 44"
-        ></path>
-        <path
-          fill="#1976D2"
-          d="M43.611 20.083H42V20H24v8h11.303a12.04 12.04 0 0 1-4.087 5.571l.003-.002 6.19 5.238C36.971 39.205 44 34 44 24c0-1.341-.138-2.65-.389-3.917"
-        ></path>
-      </svg>
-      <span>Sign in with Google</span>
-    </button>
-  );
+        Sign out
+      </button>
+    );
+  }
+
+  return <div id="googleSignInButton"></div>;
 };
 
 export default GoogleButton;
